+++ tests/frontend/org/voltdb/benchmark/tpcc/TPCCClient.java
@@ -206,10 +206,10 @@ implements TPCCSimulation.ProcCaller {
         // Order Status table 2
         constraint1 = Verification.compareWithConstant(ExpressionType.COMPARE_GREATERTHAN,
                                                        "OL_SUPPLY_W_ID",
-                                                       (byte) 0);
+                                                       (short) 0);
         constraint2 = Verification.compareWithConstant(ExpressionType.COMPARE_LESSTHANOREQUALTO,
                                                        "OL_SUPPLY_W_ID",
-                                                       (byte) m_scaleParams.warehouses);
+                                                       (short) m_scaleParams.warehouses);
         constraint3 = Verification.compareWithConstant(ExpressionType.COMPARE_GREATERTHANOREQUALTO,
                                                        "OL_I_ID",
                                                        0);
@@ -323,7 +323,7 @@ implements TPCCSimulation.ProcCaller {
     }
 
     @Override
-    public void callDelivery(byte w_id, int carrier, TimestampType date) throws IOException {
+    public void callDelivery(short w_id, int carrier, TimestampType date) throws IOException {
         m_queuedMessage = m_voltClient.callProcedure(new DeliveryCallback(),
                 Constants.DELIVERY, w_id, carrier, date);
     }
@@ -404,8 +404,9 @@ implements TPCCSimulation.ProcCaller {
 
 
     @Override
-    public void callPaymentById(byte w_id, byte d_id, double h_amount,
-            byte c_w_id, byte c_d_id, int c_id, TimestampType now) throws IOException {
+    public void callPaymentById(short w_id, byte d_id, double h_amount,
+            short c_w_id, byte c_d_id, int c_id, TimestampType now) throws IOException {
+
        if (m_scaleParams.warehouses > 1) {
            m_voltClient.callProcedure(new VerifyBasicCallback(),
                    Constants.PAYMENT_BY_ID_W,
@@ -424,8 +425,8 @@ implements TPCCSimulation.ProcCaller {
     }
 
     @Override
-    public void callPaymentByName(byte w_id, byte d_id, double h_amount,
-            byte c_w_id, byte c_d_id, byte[] c_last, TimestampType now) throws IOException {
+    public void callPaymentByName(short w_id, byte d_id, double h_amount,
+            short c_w_id, byte c_d_id, byte[] c_last, TimestampType now) throws IOException {
         if ((m_scaleParams.warehouses > 1) || (c_last != null)) {
             m_voltClient.callProcedure(new VerifyBasicCallback(),
                     Constants.PAYMENT_BY_NAME_W, w_id, d_id, h_amount,
@@ -457,7 +458,7 @@ implements TPCCSimulation.ProcCaller {
       }
 
     @Override
-    public void callStockLevel(byte w_id, byte d_id, int threshold) throws IOException {
+    public void callStockLevel(short w_id, byte d_id, int threshold) throws IOException {
         m_queuedMessage = m_voltClient.callProcedure(new StockLevelCallback(), Constants.STOCK_LEVEL,
                 w_id, d_id, threshold);
